workflows: update job permissions for test reporting
Test reporting action requires the following permissions:

  checks: write
  pull-requests: write

This patch adds the required permissions to all jobs that require test
reporting. In addition the following permissions are defined:

  contents: read
  packages: read

These are default values set by github, but need to be manually defined
to prevent setting them to "none" as per actions documentation.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@oss.qualcomm.com>