QCLINUX: networkmanager: Specify a port for NetworkManager dnsmasq
When user creates a shared hotspot through nmcli, NetworkManager
will start a private dnsmasq service.

If the system already has a global dnsmasq service and is using
the default port 53. The private dnsmasq launched by NetworkManager
may conflict with it while creating listening socket with default
port.

So it is necessary to specify the port of the NetworkManager dnsmasq
to avoid this private dnsmasq using default port.

Change-Id: Ia37711a67c0f0c490007adadc1ba61be89e5008d
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>