linux-qcom-base-6.6: Disable auto_off when BT_EN is pulled up by hw
If the BT SoC BT_EN is controlled by hardware, disable the AUTO_OFF
feature. Otherwise, BT will close the HCI layer except for the UART
after firmware download. However, the SoC remains active. If the SoC
sends a packet to the Host after firmware download, the Host cannot
respond since the HCI layer is closed, which will cause the firmware
to enter an incorrect state.

Change-Id: I9dad68b8956cee92ac7ec711ff2a10e27884d6f5
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>