linux-qcom-base-6.6: SSR(Sub-System Restart)process failed due to tx_idle_timer timeout
When the SSR (Sub-System Restart) duration exceeds 2 seconds, it triggers
host tx_idle_timeout, which sets host TX state to sleep. due to the
hardware pulling up bt_en, the firmware is not downloaded after the SSR.
As a result, the controller does not enter sleep mode. Consequently,
when the host sends a command afterward, it sends 0xFD to the controller,
but the controller does not respond, leading to a command timeout.

So reset tx_idle_timer after SSR to prevent host enter TX IBS_Sloeep mode.

Change-Id: I7531d944d7619dbaeff6170fa6dd085c8c65b326
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>