linux-qcom-base-6.6: update the high bitfield of certain DSI registers
Currently, the high bitfield of certain DSI registers
do not align with the configuration of the SWI registers
description. This can lead to wrong programming these DSI
registers, for example for 4k resolution where H_TOTAL is
taking 13 bits but software is programming only 12 bits
because of the incorrect bitmask for H_TOTAL bitfield,
this is causing DSI FIFO errors. To resolve this issue,
increase the high bitfield of the DSI registers from 12 bits
to 16 bits in dsi.xml to match the SWI register configuration.

Change-Id: I3767d661afd6f7e1520661e650fcb3e1ac208d8a
Signed-off-by: Ayushi Makhija <amakhija@qti.qualcomm.com>
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>