linux-firmware: introduce bbappend to include downstream firmware
While all HLOS firmware is ideally sourced from `linux-firmware.git`,
there are cases where it may not be up to date. In such scenarios,
specific firmware components need to be picked from downstream repos.

This bbappend enables inclusion of such downstream firmware and
installs it under `/lib/firmware/updates`, allowing the kernel to
override or supplement upstream firmware as needed.

The initial set of downstream firmware added through this mechanism
is for the GPU.

Change-Id: I6aa293ca4280c809cc09f43dc3b64a64fce441b7
Signed-off-by: Vishwas Udupa <vudupa@qti.qualcomm.com>
Signed-off-by: Viswanath Kraleti <vkraleti@qti.qualcomm.com>
Signed-off-by: Ramya Dasineni <rdasinen@qti.qualcomm.com>