QCLINUX: update size and number of files in rsyslog rotate file
There is an increase in shmem from syslog which is having an active size limit of 1GB as syslog continues to receive new message and grow even after journald stops growing. So to decrease the shmem, reduce the active log size to 100MB and increase the number of files to 20.

Change-Id: Iab3c232af39b5631324941e7b67561c248106e4a
Signed-off-by: Ramya Dasineni <rdasinen@qti.qualcomm.com>