linux-qcom-base_6.6.bb: Update the SDCC clock to use shared_floor_ops
Fix "gcc_sdcc2_apps_clk_src: rcg didn't update its configuration" during
boot. This happens due to the floor_ops tries to update the rcg
configuration even if the clock is not enabled.
The shared_floor_ops ensures that the RCG is safely parked and the new
parent configuration is cached in the parked_cfg when the clock is off.

Ensure to use the ops for the other SDCC clock instances as well.

Fixes: 39d6dcf67fe9 ("clk: qcom: gcc: Add support for QCS615 GCC clocks")

Change-Id: I7cf8403f868f693687b498ff181d60dd3d0bc7b0
Signed-off-by: Taniya Das <taniya.das@oss.qualcomm.com>
Signed-off-by: Dongfang Zhao <dongfang.zhao@oss.qualcomm.com>
Signed-off-by: Ramya Dasineni <rdasinen@qti.qualcomm.com>