PENDING: misc: fastrpc: Return on argument copy failure
Patch for copying data from kernel to user space fails, the current code
still tries to call the DSP and release resources. But if unmapping
those resources fails, it can lead to resoursce leak and potential
fault issues.

Avoid this by not calling unmap and free in case of failure and rely
on process exit for cleanup.

Change-Id: I697f779a8a647be98de530b4da0346e6acd0ec9c
Signed-off-by: Vamsi Krishna Putti <vputti@qti.qualcomm.com>